Move Word.bulk_create() to make_many_words()
Also moved the Puzzle type alias from Game to Utils.
